View holder with item updated

+++ app/src/main/java/com/example/currencyapp/ui/ExchangeRate.kt
@@ -2,5 +2,5 @@ package com.example.currencyapp.ui
 
 data class ExchangeRate(
     val name: String,
-    val rate: Long
+    val rate: Double
 )
